Update install.sh to not overwrite existing files with custom settings. Fix bug in panel_preset.py that caused long prefixes of '# # #..' to appear on comment lines. Change katana.py to duck volume only once with 200ms delays.  Update virt_start.sh to agree with my current development machine. YMMV.

+++ install.sh
@@ -9,13 +9,19 @@ UDEVDIR=/etc/udev/rules.d
 if ! `grep -q katana-user /etc/passwd`; then
     echo "Create non-privileged user for MIDI bridge"
     useradd -m -s /bin/false -G plugdev,audio katana-user
+else
+    echo "Katana user already exists"
 fi
 
 echo "Copy program and support scripts to $BINDIR"
 
-cp -f katana_bridge_start $BINDIR
-chmod 0755 $BINDIR/katana_bridge_start
-chown root:root $BINDIR/katana_bridge_start
+if [ -f "$BINDIR/katana_bridge_start" ]; then
+    echo "Not overwriting existing katana_bridge_start"
+else
+    cp -f katana_bridge_start $BINDIR
+    chmod 0755 $BINDIR/katana_bridge_start
+    chown root:root $BINDIR/katana_bridge_start
+fi
 
 cp -f katana_bridge_stop $BINDIR
 chmod 0755 $BINDIR/katana_bridge_stop
@@ -26,10 +32,10 @@ chmod 0755 $BINDIR/katana_bridge_app
 chown root:root $BINDIR/katana_bridge_app
 
 echo "Copy Python modules and parameter files to $LIBDIR"
-mkdir -p $LIBDIR
+[ -d $LIBDIR ] || mkdir -p $LIBDIR
 cp -f *.py $LIBDIR
 
-mkdir -p $PARMDIR
+[ -d $PARMDIR ] || mkdir -p $PARMDIR
 cp -f ./parameters/*.json $PARMDIR
 
 chown -R root:root $LIBDIR/*
@@ -46,13 +52,21 @@ $INITDIR/katana_bridge start
 
 echo "Copy udev rules to $UDEVDIR and refresh system"
 
-cp -f 50-katana.rules $UDEVDIR
-chmod 0644 $UDEVDIR/50-katana.rules
-chown root:root $UDEVDIR/50-katana.rules
+if [ -f "$UDEVDIR/50-katana.rules" ]; then
+    echo "Not overwriting existing 50-katana.rules"
+else
+    cp -f 50-katana.rules $UDEVDIR
+    chmod 0644 $UDEVDIR/50-katana.rules
+    chown root:root $UDEVDIR/50-katana.rules
+fi
 
-cp -f 60-controller.rules $UDEVDIR
-chmod 0644 $UDEVDIR/60-controller.rules
-chown root:root $UDEVDIR/60-controller.rules
+if [ -f "$UDEVDIR/60-controller.rules" ]; then
+    echo "Not overwriting existing 60-controller.rules"
+else
+    cp -f 60-controller.rules $UDEVDIR
+    chmod 0644 $UDEVDIR/60-controller.rules
+    chown root:root $UDEVDIR/60-controller.rules
+fi
 
 udevadm control --reload
